Improved when/how modal parses URL's and displays errors
Now keeps track of the current URL, so if it doesn’t change at all but
a change event gets fired, it won’t re-parse. If the URL field is
blank, it won’t re-parse, and just resets the form back to the original
state. Errors are removed and the media link is disabled after a URL
change event occurs.